Update libloragw-sx1302.
This includes a fix for RAK shields that do not have a temperature
sensor. If no temp. sensor is found, the i2c descriptor is closed and
the descriptor is set to -1. On lgw_stop the i2c descriptor was closed
again. This caused the Concentratord to panic in case of configuration
update.