Update ChirpStack components.
As a temp workaround, this patches the go.mod and go.sum of the
ChirpStack components, as 'go 1.18' does break with the Go compiler
version that is included in this Yocto version.

The main purpose for this update is to update the ChirpStack Gateway
Bridge to v3.14.x, which provides compatibility with ChirpStack v4.