Include target config + disable console on UART.
This migrates the chirpstack-openwrt-config repository content into this
repository and updates the Makefile for switching configuration
environments.

The reason for this is that ./scripts/env does handle .config + files
per environment, but it does not make it possible to apply patches. To
make sure that we are not sending console output to a GNSS module
connected to the UART interface, we need to patch the cmdline.txt.

The updated Makefile handles the config switching in the same way as the
./scripts/env would do, but also handles the rollback and apply of
patches on switching the configuration environment.