rak7268v2: Fix random wifi mac + hostname.
This adds a hotplug script that is triggered after creating a new wlan
interface. It will read the mac from the factory partition and increment
by 1 (to not collide with the ethernet mac).

Instead of basing the hostname on the mac of eth0, it also reads from
the factory partition.