rpi: Add Raspberry Pi 5 full-image config.
Notes:

This contains a patch to add additional kernel modules to enable the
SPI interface of the Raspberry Pi 5. This patch enables the following
modules (other changes were automatically introduced):

CONFIG_SPI_DESIGNWARE=y
CONFIG_SPI_DW_MMIO=y

This also contains the following Redis configuration:

ignore-warnings ARM64-COW-BUG

This is needed as in the past there was an issue that could cause
data-corruption. This has been fixed in the Linux kernel, but as Redis
is not able to detect if the change has implemented or not, it will not
start without the above configuration.

See also #128.